E-Commerce-Chatbot/requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
flask
langchain_google_genai
-e .